django_projects/api_geek/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
dj-database-url==2.3.0
django-filter==24.3
django-picklefield==3.2
django-polymorphic==3.1.0
django-q==1.3.9
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
drf-yasg==1.21.8
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
whitenoise==6.9.0
